fixes[t:535]

+++ engine/core/audio/soundemitter.cpp
@@ -125,7 +125,7 @@ namespace FIFE {
 		CHECK_OPENAL_LOG(_log, LogManager::LEVEL_ERROR, "error attaching sound clip")
 	}
 
-	void SoundEmitter::updateEvent(uint64_t time) {
+	void SoundEmitter::updateEvent(uint32_t time) {
 		ALint procs;
 		ALint bufs;
 		ALuint buffer;
