+++ engine/core/util/resource/resourcemanager.h
@@ -75,7 +75,7 @@ namespace FIFE {
 		 * @return Total resources created
 		 *
 		 */
-		virtual uint32_t getTotalResourcesCreated() const = 0;
+		virtual size_t getTotalResourcesCreated() const = 0;
 
 		/** Returns the number of loaded resources
 		 *
@@ -86,7 +86,7 @@ namespace FIFE {
 		 * @return Total resources loaded
 		 *
 		 */
-		virtual uint32_t getTotalResourcesLoaded() const = 0;
+		virtual size_t getTotalResourcesLoaded() const = 0;
 
 		/** Returns the number of defined resources
 		 *
@@ -95,7 +95,7 @@ namespace FIFE {
 		 * @return Total resources managed
 		 *
 		 */
-		virtual uint32_t getTotalResources() const = 0;
+		virtual size_t getTotalResources() const = 0;
 	};
 
 } //FIFE
