+++ engine/core/view/renderers/genericrenderer.cpp
@@ -150,8 +150,8 @@ namespace FIFE {
 
 			Rect r;
 			Rect viewport = cam->getViewPort();
-			uint32_t widtht = round(img->getWidth() * cam->getZoom());
-			uint32_t height = round(img->getHeight() * cam->getZoom());
+			uint32_t widtht = static_cast<uint32_t>(round(img->getWidth() * cam->getZoom()));
+			uint32_t height = static_cast<uint32_t>(round(img->getHeight() * cam->getZoom()));
 			r.x = p.x-widtht/2;
 			r.y = p.y-height/2;
 			r.w = widtht;
@@ -180,8 +180,8 @@ namespace FIFE {
 			ImagePtr img = ImageManager::instance()->get(animation.getFrameByTimestamp(animtime));
 			Rect r;
 			Rect viewport = cam->getViewPort();
-			uint32_t widtht = round(img->getWidth() * cam->getZoom());
-			uint32_t height = round(img->getHeight() * cam->getZoom());
+			uint32_t widtht = static_cast<uint32_t>(round(img->getWidth() * cam->getZoom()));
+			uint32_t height = static_cast<uint32_t>(round(img->getHeight() * cam->getZoom()));
 			r.x = p.x-widtht/2;
 			r.y = p.y-height/2;
 			r.w = widtht;
@@ -233,8 +233,8 @@ namespace FIFE {
 
 			Rect r;
 			Rect viewport = cam->getViewPort();
-			uint32_t widtht = round(m_width * cam->getZoom());
-			uint32_t height = round(m_height * cam->getZoom());
+			uint32_t widtht = static_cast<uint32_t>(round(m_width * cam->getZoom()));
+			uint32_t height = static_cast<uint32_t>(round(m_height * cam->getZoom()));
 			r.x = p.x-widtht/2;
 			r.y = p.y-height/2;
 			r.w = widtht;
